+++ Frontend/scripts/studentPage/links.js
@@ -6,6 +6,8 @@ import { displayRequestPage } from "./displayRequestPage.js";
 import { clearSelected } from "../common/clearSelected.js";
 import { getRequests } from "../common/getRequests.js";
 import { createWebSocketConnection } from "../common/createWebSocketConnection.js";
+import { logout } from "../common/logout.js";
+import { serverErrorMessage } from "../common/serverErrorMessage.js";
 
 // Display STATUS page as default
 let requests = await getRequests();
@@ -25,7 +27,7 @@ document.body.addEventListener('updateRequests', async (event) => {
 });
 
 // Define links for each page
-document.querySelector('.js-sidebar-container').addEventListener('click', (event) => {
+document.querySelector('.js-sidebar-container').addEventListener('click', async (event) => {
     if (event.target.classList.contains('js-status-link')) {
         clearSelected();
         event.target.classList.add('selected');
@@ -35,8 +37,12 @@ document.querySelector('.js-sidebar-container').addEventListener('click', (event
         event.target.classList.add('selected');
         displayRequestPage();
     } else if (event.target.classList.contains('js-logout-link')) {
-        sessionStorage.clear();
-        window.location.href = 'studentLogin.html';
+        if (await logout()) {           
+            sessionStorage.clear();
+            window.location.href = 'studentLogin.html';
+        } else {
+            serverErrorMessage();
+        }
     }
 });
